Issue #30: Use Arc in dispatcher
When you clone an Event, you're cloning all of the data. If the event is
non-trivial, that's a lot of cloned memory. If you use an Arc, you can
clone the Arc, and each thread gets his own Arc, and each thread
references the same memory without having to clone that memory.